Revert "chore: Created new controller for project transfer"
This reverts commit 74e107ade5ced8ee62205bea9e1613bd308b8980.
